requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
attrs==21.2.0
    # via dkpro-cassis
deprecation==2.1.0
    # via dkpro-cassis
dkpro-cassis==0.8.0
    # via -r requirements.in
importlib-resources==5.4.0
    # via dkpro-cassis
lxml==4.9.4
    # via dkpro-cassis
more-itertools==8.12.0
    # via dkpro-cassis
packaging==23.2
    # via deprecation
sortedcontainers==2.4.0
    # via dkpro-cassis
toposort==1.7
    # via dkpro-cassis
matplotlib~=3.8.3
python-bidi
arabic-reshaper
    # via -r requirements.i
numpy~=1.24.3
scikit-learn~=1.3.2
pandas~=2.1.1
requests~=2.31.0
sympy~=1.12
Levenshtein~=0.25.0